Fixes for double to int cast warnings. ASSIGNED - bug 1401: Use vertex arrays with OpenGL https://tracker.physiomeproject.org/show_bug.cgi?id=1401
git-svn-id: https://svn.physiomeproject.org/svn/cmiss/cmgui/trunk@6057 3c1deb5b-d424-0410-962d-aba41a686d42

+++ source/graphics/scene.cpp
@@ -3968,7 +3968,7 @@ DESCRIPTION :
 Returns 1 if the Scene object has a time dependence, 0 if not
 ==============================================================================*/
 {
-	double return_code;
+	int return_code;
 
 	ENTER(Scene_object_has_time);
 	if (scene_object)
